제곱수 판별하기.java

+++ programmers/0레벨/제곱수 판별하기.java
@@ -3,7 +3,7 @@ import java.util.*;
 class Solution {
     public int solution(int n) {
         int answer = 0;
-        double a = Math.sqrt(n);
+        int a = (int)Math.sqrt(n);
         
         if(a*a==n){
             answer=1;
